// src/interfaces/forecastsFiveDays.ts
  export interface Headline {
    EffectiveDate: string;
    EffectiveEpochDate: number;
    Severity: number;
    Text: string;
    Category: string;
    EndDate?: any;
    EndEpochDate?: any;
    MobileLink: string;
    Link: string;
  }

  export interface Minimum {
    Value: number;
    Unit: string;
    UnitType: number;
  }

  export interface Maximum {
    Value: number;
    Unit: string;
    UnitType: number;
  }

  export interface Temperature {
    Minimum: Minimum;
    Maximum: Maximum;
  }

  export interface Day {
    Icon: number;
    IconPhrase: string;
    HasPrecipitation: boolean;
  }

  export interface Night {
    Icon: number;
    IconPhrase: string;
    HasPrecipitation: boolean;
  }

  export interface DailyForecast {
    Date: string;
    EpochDate: number;
    Temperature: Temperature;
    Day: Day;
    Night: Night;
    Sources: string[];
    MobileLink: string;
    Link: string;
  }

  export interface IForceCastsFiveDays {
    Headline: Headline;
    DailyForecasts: DailyForecast[];
  }